Move SetCommonConditions into generic
