+++ internal/controllers/routerinterface/status.go
@@ -27,7 +27,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	orcv1alpha1 "github.com/k-orc/openstack-resource-controller/api/v1alpha1"
-	"github.com/k-orc/openstack-resource-controller/internal/controllers/common"
+	"github.com/k-orc/openstack-resource-controller/internal/controllers/generic"
 	"github.com/k-orc/openstack-resource-controller/internal/util/applyconfigs"
 	orcapplyconfigv1alpha1 "github.com/k-orc/openstack-resource-controller/pkg/clients/applyconfiguration/api/v1alpha1"
 )
@@ -80,7 +80,7 @@ func createStatusUpdate(orcObject *orcv1alpha1.RouterInterface, now metav1.Time,
 	}
 
 	isAvailable, isUpToDate, progressMessage := getStatusSummary(orcObject, statusOpts)
-	common.SetCommonConditions(orcObject, applyConfigStatus, isAvailable, isUpToDate, progressMessage, statusOpts.err, now)
+	generic.SetCommonConditions(orcObject, applyConfigStatus, isAvailable, isUpToDate, progressMessage, statusOpts.err, now)
 
 	return applyConfig
 }
